+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## to be released
 
-- enable Travis CI
+- enable Travis CI and SensioLabsInsight
 - add shims for mb_check_encoding, mb_detect_encoding, mb_detect_order,
   mb_language and mb_encoding_aliases
 - mbstring shim fix: alias UTF8 to UTF-8