+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## to be released
+## v1.1.17 (2014-01-02)
 
 - enable Travis CI and SensioLabsInsight
 - add shims for mb_check_encoding, mb_detect_encoding, mb_detect_order,
   mb_language and mb_encoding_aliases
 - mbstring shim fix: alias UTF8 to UTF-8
+- more tests
 
 ## v1.1.16 (2013-12-06)